+++ packages/helpers/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @swapkit/helpers
 
+## 4.0.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/tokens@4.0.20
+
+## 4.0.19
+
+### Patch Changes
+
+- [#1494](https://github.com/swapkit/SwapKit/pull/1494) [`6365b50`](https://github.com/swapkit/SwapKit/commit/6365b506e3ed6587298fa6ba71c846bdf7341391) Thanks [@towanTG](https://github.com/towanTG)! - Corrects block time for zcash and doge
+
+- Updated dependencies []:
+  - @swapkit/tokens@4.0.19
+
+## 4.0.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/tokens@4.0.18
+
+## 4.0.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/tokens@4.0.17
+
+## 4.0.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/tokens@4.0.16
+
+## 4.0.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/tokens@4.0.15
+
 ## 4.0.14
 
 ### Patch Changes