+++ packages/wallet-core/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @swapkit/wallet-core
 
+## 4.0.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/helpers@4.0.20
+
+## 4.0.19
+
+### Patch Changes
+
+- Updated dependencies [[`6365b50`](https://github.com/swapkit/SwapKit/commit/6365b506e3ed6587298fa6ba71c846bdf7341391)]:
+  - @swapkit/helpers@4.0.19
+
+## 4.0.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/helpers@4.0.18
+
+## 4.0.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/helpers@4.0.17
+
+## 4.0.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/helpers@4.0.16
+
+## 4.0.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/helpers@4.0.15
+
 ## 4.0.14
 
 ### Patch Changes